Replace ellipses with three full stops
As much as this pains me we can't use a literal ellipsis character
because some tools barf when they encounter it, printing question marks
instead. Obviously questioning whether something is starting is not
ideal.

> Starting Application?

We can't fix all the tools out there so I suggest it's better we use
a more portable, albeit less accurate set of glyphs.